+++ src/main/java/ar/com/plug/examen/domain/exception/ProductNotFoundException.java
@@ -1,10 +1,9 @@
 package ar.com.plug.examen.domain.exception;
 
 public class ProductNotFoundException extends RuntimeException {
-
 	private static final long serialVersionUID = 3095543865324828748L;
 
-	public ProductNotFoundException(Integer id) {
+	public ProductNotFoundException(Long id) {
 		super("Cannot find Product " + id);
 	}
 }
